Dose response for Physical Activity

Based on our meta-analysis, returns a response value for a specific cause or disease.
1. Disease/Cause: Name of the disease/cause - from a specific list
2. Type: Fatal, non-fatal or fatal-and-non-fatal.

dose_response(
  cause,
  outcome_type,
  dose,
  quantile = 0.5,
  censor_method = "default",
  confidence_intervals = F
)

Arguments

cause

All-cause-mortality or disease name

outcome_type

Fatal, non-fatal or fatal-and-non-fatal

dose

Dose (all-cause or disease)

quantile

Numeric value between 0 and 1 - default is 0.5

censor_method

Use either no censor method, or the 75th percentile by person years,

confidence_intervals

Boolean variable to determine whether confidence intervals are returned or not - default is FALSE

Value

response for a specific dose (in a data frame)

Details

At the end, this function returns a response values with or without confidence interval